Refactored parallel.c to use smaller functions in parallel_zip, as well as fixed significant memory leaks from forgetting to free larger buffers. As it stands, all buffers are freed through free_buffer_pool, but this is only reliable because of careful convention to always ensure that all buffers ever allocated are present in one of the pools at the end of the run. In the future, we might consider establishing a much more rigorous framework to ensure that this is invariably the case, or some other method to both free all malloc'd memory and avoid double frees, as both job_list and buffer_pool have pointers into buffers, and invalidating one does not invalidate the other, leading to difficulties freeing.

+++ src/gzip.h
@@ -36,6 +36,7 @@
  */
 #include <stdio.h>
 #include <sys/types.h> /* for off_t */
+#include <stdint.h>
 #include <time.h>
 #include <string.h>
 #include <stdnoreturn.h>
@@ -201,12 +202,12 @@ typedef int file_t;     /* Do not use stdio */
 extern int exit_code;      /* program exit code */
 extern int verbose;        /* be verbose (-v) */
 extern int quiet;          /* be quiet (-q) */
-extern int level;          /* compression level */
+extern uint8_t level;          /* compression level */
 extern int test;           /* check .z file integrity */
 extern int to_stdout;      /* output to stdout (-c) */
 extern int save_orig_name; /* set if original name must be saved */
 extern int no_name;        /* original name should not be restored */
-extern int pkzip;          /* global variable for pkzip */ 
+extern int pkzip;          /* global variable for pkzip */
 
 #define get_byte()  (inptr < insize ? inbuf[inptr++] : fill_inbuf(false, CHUNK))
 #define try_byte()  (inptr < insize ? inbuf[inptr++] : fill_inbuf(true, CHUNK))
